Added changes as suggested by @truls
