+++ test/Accesslist.events.js
@@ -1,6 +1,7 @@
+const events = require('./events.js');
 const utils = require('./utils.js');
 
-module.exports = utils.makeEventMap({
+module.exports = events.makeEventMap({
   // TODO: Implement the constructor using an inheritance system
   constructor: (addr) => [
     { eventName: 'OwnershipTransferred',
