+++ test/token/ERC20/ExternalERC20Mintable.events.js
@@ -1,6 +1,6 @@
-const utils = require('./../..//utils.js');
+const events = require('./../../events.js');
 
-module.exports = utils.makeEventMap({
+module.exports = events.makeEventMap({
   changeMintingRecipient: (prevAddr, nextAddr) => [
     { eventName: 'MintingRecipientAccountChanged',
       paramMap: { prev: prevAddr,
